+++ packages/decap-cms-core/index.d.ts
@@ -10,6 +10,7 @@ declare module 'decap-cms-core' {
     | 'github'
     | 'gitlab'
     | 'gitea'
+    | 'forgejo'
     | 'bitbucket'
     | 'test-repo'
     | 'proxy';
@@ -60,6 +61,7 @@ declare module 'decap-cms-core' {
     i18n?: boolean | 'translate' | 'duplicate' | 'none';
     media_folder?: string;
     public_folder?: string;
+    media_processing?: CmsMediaProcessing;
     comment?: string;
   }
 
@@ -365,6 +367,7 @@ declare module 'decap-cms-core' {
       editor?: { preview?: boolean };
       fields?: CmsField[];
     };
+    limit?: number;
 
     /**
      * @deprecated Use sortable_fields instead
@@ -414,6 +417,21 @@ declare module 'decap-cms-core' {
     url?: string;
   }
 
+  export type CmsMediaProcessingFormat = 'jpeg' | 'webp';
+
+  export interface CmsMediaProcessing {
+    enabled: boolean;
+    format?: {
+      enabled: boolean;
+      default: CmsMediaProcessingFormat;
+    };
+    quality?: number;
+    strip_metadata?: boolean;
+    width?: number | null;
+    height?: number | null;
+    aspect_ratio?: number | string | null;
+  }
+
   export interface CmsConfig {
     backend: CmsBackend;
     collections: CmsCollection[];
@@ -429,6 +447,7 @@ declare module 'decap-cms-core' {
     media_folder?: string;
     public_folder?: string;
     media_folder_relative?: boolean;
+    media_processing?: CmsMediaProcessing;
     media_library?: CmsMediaLibrary;
     publish_mode?: CmsPublishMode;
     issue_reports?: CmsIssueReports;