+++ packages/decap-cms-core/src/components/Collection/Collection.js
@@ -1,4 +1,5 @@
-import React from 'react';
+// eslint-disable-next-line no-unused-vars
+import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import ImmutablePropTypes from 'react-immutable-proptypes';
 import styled from '@emotion/styled';
@@ -24,6 +25,7 @@ import {
   selectEntriesGroup,
   selectViewStyle,
 } from '../../reducers/entries';
+import { selectCanCreateNewEntry } from '../../reducers';
 
 const CollectionContainer = styled.div`
   display: flex;
@@ -49,7 +51,7 @@ const SearchResultHeading = styled.h1`
   ${components.cardTopHeading};
 `;
 
-export class Collection extends React.Component {
+export class Collection extends Component {
   static propTypes = {
     searchTerm: PropTypes.string,
     collectionName: PropTypes.string,
@@ -60,6 +62,9 @@ export class Collection extends React.Component {
     sortableFields: PropTypes.array,
     sort: ImmutablePropTypes.orderedMap,
     onSortClick: PropTypes.func.isRequired,
+    canCreate: PropTypes.bool.isRequired,
+    filterTerm: PropTypes.string,
+    viewStyle: PropTypes.string,
   };
 
   componentDidMount() {
@@ -106,9 +111,10 @@ export class Collection extends React.Component {
       group,
       onChangeViewStyle,
       viewStyle,
+      canCreate,
     } = this.props;
 
-    let newEntryUrl = collection.get('create') ? getNewEntryUrl(collectionName) : '';
+    let newEntryUrl = canCreate ? getNewEntryUrl(collectionName) : '';
     if (newEntryUrl && filterTerm) {
       newEntryUrl = getNewEntryUrl(collectionName);
       if (filterTerm) {
@@ -174,6 +180,7 @@ function mapStateToProps(state, ownProps) {
   const filter = selectEntriesFilter(state.entries, collection.get('name'));
   const group = selectEntriesGroup(state.entries, collection.get('name'));
   const viewStyle = selectViewStyle(state.entries);
+  const canCreate = selectCanCreateNewEntry(state, name);
 
   return {
     collection,
@@ -190,6 +197,7 @@ function mapStateToProps(state, ownProps) {
     filter,
     group,
     viewStyle,
+    canCreate,
   };
 }