+++ packages/decap-cms-core/src/components/Editor/Editor.js
@@ -1,5 +1,5 @@
+import { Component } from 'react';
 import PropTypes from 'prop-types';
-import React from 'react';
 import ImmutablePropTypes from 'react-immutable-proptypes';
 import { connect } from 'react-redux';
 import { Loader } from 'decap-cms-ui-default';
@@ -36,7 +36,12 @@ import {
   deleteUnpublishedEntry,
 } from '../../actions/editorialWorkflow';
 import { loadDeployPreview } from '../../actions/deploys';
-import { selectEntry, selectUnpublishedEntry, selectDeployPreview } from '../../reducers';
+import {
+  selectCanCreateNewEntry,
+  selectEntry,
+  selectUnpublishedEntry,
+  selectDeployPreview,
+} from '../../reducers';
 import { selectFields, getFileFromSlug } from '../../reducers/collections';
 import { status, EDITORIAL_WORKFLOW } from '../../constants/publishModes';
 import { FILES } from '../../constants/collectionTypes';
@@ -53,7 +58,7 @@ function isNotesEnabled(collection, slug) {
   return collection.getIn(['editor', 'notes'], false);
 }
 
-export class Editor extends React.Component {
+export class Editor extends Component {
   static propTypes = {
     changeDraftField: PropTypes.func.isRequired,
     changeDraftFieldValidation: PropTypes.func.isRequired,
@@ -92,6 +97,7 @@ export class Editor extends React.Component {
       search: PropTypes.string,
     }),
     hasChanged: PropTypes.bool,
+    canCreateNewEntry: PropTypes.bool,
     t: PropTypes.func.isRequired,
     retrieveLocalBackup: PropTypes.func.isRequired,
     localBackup: ImmutablePropTypes.map,
@@ -401,6 +407,7 @@ export class Editor extends React.Component {
       changeDraftFieldValidation,
       user,
       hasChanged,
+      canCreateNewEntry,
       displayUrl,
       hasWorkflow,
       useOpenAuthoring,
@@ -455,6 +462,7 @@ export class Editor extends React.Component {
         showDelete={this.props.showDelete}
         user={user}
         hasChanged={hasChanged}
+        canCreateNewEntry={canCreateNewEntry}
         displayUrl={displayUrl}
         hasWorkflow={hasWorkflow}
         useOpenAuthoring={useOpenAuthoring}
@@ -487,6 +495,7 @@ function mapStateToProps(state, ownProps) {
   const entry = newEntry ? null : selectEntry(state, collectionName, slug);
   const user = auth.user;
   const hasChanged = entryDraft.get('hasChanged');
+  const canCreateNewEntry = selectCanCreateNewEntry(state, collectionName);
   const displayUrl = config.display_url;
   const hasWorkflow = config.publish_mode === EDITORIAL_WORKFLOW;
   const useOpenAuthoring = globalUI.useOpenAuthoring;
@@ -520,6 +529,7 @@ function mapStateToProps(state, ownProps) {
     entry,
     user,
     hasChanged,
+    canCreateNewEntry,
     displayUrl,
     hasWorkflow,
     useOpenAuthoring,