+++ packages/decap-cms-ui-default/src/styles.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import { css, Global } from '@emotion/react';
 
 /**