+++ packages/decap-cms-app/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.16.0](https://github.com/decaporg/decap-cms/compare/decap-cms-app@3.15.1...decap-cms-app@3.16.0) (2026-08-31)
+
+**Note:** Version bump only for package decap-cms-app
+
+## [3.15.1](https://github.com/decaporg/decap-cms/compare/decap-cms-app@3.15.0...decap-cms-app@3.15.1) (2026-07-24)
+
+**Note:** Version bump only for package decap-cms-app
+
+# [3.15.0](https://github.com/decaporg/decap-cms/compare/decap-cms-app@3.14.1...decap-cms-app@3.15.0) (2026-07-23)
+
+### Features
+
+- add uuid widget ([#6675](https://github.com/decaporg/decap-cms/issues/6675)) ([78c0793](https://github.com/decaporg/decap-cms/commit/78c079313e0eb2f0f98d2734338ead8046dfdba4))
+
 ## [3.14.1](https://github.com/decaporg/decap-cms/compare/decap-cms-app@3.14.0...decap-cms-app@3.14.1) (2026-06-15)
 
 **Note:** Version bump only for package decap-cms-app